chore: move generateTestMatrix.js and its tests under src
